add ABI check, improve C++ code and output check

+++ call_c.c
@@ -1,4 +1,4 @@
-void timestwo(double x[], double x2[], int N){
+void timestwo(int x[], int x2[], int N){
   for (int i=0; i<N; i++)
     x2[i] = x[i] * 2;
 }
