+++ call_cxx.cxx
@@ -3,7 +3,7 @@
 // nm libcxx.a
 // or whatever you name this library file.
 
-extern "C" void timestwo(double x[], double x2[], int N){
+extern "C" void timestwo(int x[], int x2[], int N){
   for (auto i=0; i<N; i++)
     x2[i] = x[i] * 2;
 }
